[fix] id가 long 에서 int로 변환에따라 사용했던 부분들 수정, jpa.hibernate.ddl-auto를 validate로 변경

+++ src/main/java/com/fab/banggabgo/service/ArticleService.java
@@ -4,27 +4,26 @@ import com.fab.banggabgo.dto.ArticleEditDto;
 import com.fab.banggabgo.dto.ArticlePageDto;
 import com.fab.banggabgo.dto.ArticleRegisterDto;
 import java.util.List;
-import org.springframework.data.domain.Pageable;
 
 public interface ArticleService {
 
-  /**
-   * 게시글 등록
-   */
-  void postArticle(String token, ArticleRegisterDto dto);
+    /**
+     * 게시글 등록
+     */
+    void postArticle(String token, ArticleRegisterDto dto);
 
-  /**
-   * 게시글 수정
-   */
-  void putArticle(String token, Long id, ArticleEditDto dto);
+    /**
+     * 게시글 수정
+     */
+    void putArticle(String token, Integer id, ArticleEditDto dto);
 
-  /**
-   * 게시글 삭제
-   */
-  void deleteArticle(String token, Long id);
+    /**
+     * 게시글 삭제
+     */
+    void deleteArticle(String token, Integer id);
 
-  /**
-   * 게시글 페이징 불러오기
-   */
-  List<ArticlePageDto> getArticleByPageable(Integer page, Integer size, boolean isRecruiting);
+    /**
+     * 게시글 페이징 불러오기
+     */
+    List<ArticlePageDto> getArticleByPageable(Integer page, Integer size, boolean isRecruiting);
 }
